Move modules into runtime/ directory
