+++ analyzer/runtime/accounts.go
@@ -1,4 +1,4 @@
-package modules
+package runtime
 
 import (
 	"math/big"
