+++ analyzer/runtime/evm.go
@@ -1,4 +1,4 @@
-package modules
+package runtime
 
 import (
 	"context"
